Bug fix for covariance matrix issues. MIEcalculator.py split into MIEcalc3.py and reporting.py
