+++ extras.py
@@ -6,7 +6,6 @@ I/O stuff.  Some features are still experimental/buggy.
 import wx
 import wx.adv
 import os
-import snapshotPrinter
 from formbuild import MyFrame2  # for the help dialog
 import csv
 import xlrd  # for xls and xlsm
@@ -40,65 +39,6 @@ class VIEW(object):
     def __init__(self, frame):
         self.frame = frame  # need it for screenshot
 
-    def TakeScreenShot(self, temp_path):
-        """
-        Takes a screenshot of the screen at a given position and size and places
-        it in myImage.png in directory *temp_path*.
-        """
-        rect = self.frame.GetRect()
-        # see http://aspn.activestate.com/ASPN/Mail/Message/wxpython-users/3575899
-        # created by Andrea Gavana
-        # adjust widths for Linux (figured out by John Torres
-        # http://article.gmane.org/gmane.comp.python.wxpython/67327)
-    ## 	if sys.platform == 'linux2':
-    ## 		client_x, client_y = self.ClientToScreen((0, 0))
-    ## 		border_width = client_x - rect.x
-    ## 		title_bar_height = client_y - rect.y
-    ## 		rect.width += (border_width * 2)
-    ## 		rect.height += title_bar_height + border_width
-        #Create a DC for the whole screen area
-        dcScreen = wx.ScreenDC()  # all screens of extended desktop
-
-        #Create a Bitmap that will hold the screenshot image later on
-        #Note that the Bitmap must have a size big enough to hold the screenshot
-        #-1 means using the current default colour depth
-        bmp = wx.Bitmap(rect.width, rect.height)
-
-        #Create a memory DC that will be used for actually taking the screenshot
-        memDC = wx.MemoryDC()
-
-        #Tell the memory DC to use our Bitmap
-        #all drawing action on the memory DC will go to the Bitmap now
-        memDC.SelectObject(bmp)
-
-        #Blit (in this case copy) the actual screen on the memory DC
-        #and thus the Bitmap
-        memDC.Blit(0, #Copy to this X coordinate
-                0, #Copy to this Y coordinate
-                rect.width, #Copy this width
-                rect.height, #Copy this height
-                dcScreen, #From where do we copy?
-                rect.x, #What's the X offset in the original DC?
-                rect.y  #What's the Y offset in the original DC?
-                )
-
-        #Select the Bitmap out of the memory DC by selecting a new
-        #uninitialized Bitmap
-        memDC.SelectObject(wx.NullBitmap)
-
-        img = bmp.ConvertToImage()
-        fileName = os.path.join(temp_path, "myImage.png")
-        img.SaveFile(fileName, wx.BITMAP_TYPE_PNG)
-
-    def onPrint(self, temp_path):
-        """
-        Send screenshot to the printer using folder *temp_path* to hold a
-        temporary copy of the file.
-        """
-        printer = snapshotPrinter.SnapshotPrinter(temp_path)
-        printer.sendToPrinter()
-        printer.Close()
-
     def scaleImage(self, image, size):
         """
         Takes an image file, *image*, resizes it to the size tuple of width and
@@ -107,12 +47,6 @@ class VIEW(object):
         """
         image_file = image
         img_org = Image.open(image_file)
-        # get the size of the original image
-        #width_org, height_org = img_org.size
-        # set the resizing factor so the aspect ratio can be retained
-        # factor > 1.0 increases size
-        # factor < 1.0 decreases size
-        #factor = scale
         width = int(size[0])
         height = int(size[1])
         # best down-sizing filter
